handle coercion of arrays
