+++ src/JsonSchema/Constraints/ObjectConstraint.php
@@ -130,94 +130,36 @@ class ObjectConstraint extends Constraint
 
             if($this->factory->getCheckMode() & Constraint::CHECK_MODE_TYPE_CAST){
                 if(!($property instanceof Constraint)) {
-					$property = $this->coerce($property, $definition);
+                    $property = $this->coerce($property, $definition);
 
-					if($this->factory->getCheckMode() & Constraint::CHECK_MODE_COERCE) {
-						if (is_object($element)) {
-							$element->{$i} = $property;
-						} else {
-							$element[$i] = $property;
-						}
-					}
+                    if($this->factory->getCheckMode() & Constraint::CHECK_MODE_COERCE) {
+                        if (is_object($element)) {
+                            $element->{$i} = $property;
+                        } else {
+                            $element[$i] = $property;
+                        }
+                    }
                 }
             }
 
             if (is_object($definition)) {
                 // Undefined constraint will check for is_object() and quit if is not - so why pass it?
                 $this->checkUndefined($property, $definition, $path, $i);
-            }
-        }
-    }
+                // check array
+                if ($this->getTypeCheck()->isArray($property)) {
+                    $path = $this->incrementPath($path ?: new JsonPointer(''), $i);
+                    $this->checkArray($property, $definition, $path, $i);
 
-    /**
-     * Converts a value to boolean. For example, "true" becomes true.
-     * @param $value The value to convert to boolean
-     * @return bool|mixed
-     */
-    protected function toBoolean($value)
-    {
-        if($value === "true"){
-            return true;
-        }
-
-        if($value === "false"){
-            return false;
-        }
-
-        return $value;
-    }
-
-    /**
-     * Converts a numeric string to a number. For example, "4" becomes 4.
-     *
-     * @param mixed $value The value to convert to a number.
-     * @return int|float|mixed
-     */
-    protected function toNumber($value)
-    {
-        if(is_numeric($value)) {
-            return $value + 0; // cast to number
-        }
-
-        return $value;
-    }
-
-    protected function toInteger($value)
-    {
-        if(is_numeric($value) && (int)$value == $value) {
-            return (int)$value; // cast to number
-        }
-
-        return $value;
-    }
-
-    /**
-     * Given a value and a definition, attempts to coerce the value into the
-     * type specified by the definition's 'type' property.
-     *
-     * @param mixed $value Value to coerce.
-     * @param \stdClass $definition A definition with information about the expected type.
-     * @return bool|int|string
-     */
-    protected function coerce($value, $definition)
-    {
-        $types = isset($definition->type)?$definition->type:null;
-        if($types){
-            foreach((array)$types as $type) {
-                switch ($type) {
-                    case "boolean":
-                        $value = $this->toBoolean($value);
-                        break;
-                    case "integer":
-                        $value = $this->toInteger($value);
-                        break;
-                    case "number":
-                        $value = $this->toNumber($value);
-                        break;
+                    if (is_object($element)) {
+                        $element->{$i} = $property;
+                    } else {
+                        $element[$i] = $property;
+                    }
+                } else {
+                    $this->checkUndefined($property, $definition, $path, $i);
                 }
             }
         }
-        return $value;
     }
 
     /**
